Updated to version 20.5

+++ lib/libupnp/threadutil/src/ThreadPool.c
@@ -341,7 +341,7 @@ tp->stats.totalJobsLQ++; tp->stats.totalTimeLQ += diff; break; default:
     struct timeb t;
 
     ftime( &t );
-    srand( ( unsigned int )t.millitm + ithread_get_current_thread_id(  ) );
+    srand( ( unsigned int )t.millitm + ( unsigned int )ithread_get_current_thread_id(  ) );
     }
 
 /****************************************************************************
